+++ pytermor/common.py
@@ -21,15 +21,15 @@ logger.addHandler(logging.NullHandler())
 # logger.setLevel('DEBUG')
 ########
 
-ST = TypeVar("ST", bound=Union[str, "Renderable"])
+ST = TypeVar("ST", bound=Union[str, "IRenderable"])
 """ 
 `ST` in a method signature usually means that regular strings as well as 
-`Renderable` implementations are supported, can be intermixed, and:
+IRenderable` implementations are supported, can be intermixed, and:
 
     - return type will be *str* if and only if the type of all arguments is *str*;
     - otherwise return type will be `Text` -- *str* arguments, if any, will
-      be transformed into `Renderable` and concatenated. `Text` type is used because
-      it's the only `Renderable` that is mutable.
+      be transformed into IRenderable` and concatenated. `Text` type is used because
+      it's the only IRenderable` that is mutable.
 
 """